hm2 renamed to hw2
