--- src/main/java/edu/hm2/task3/managers/ConnectionManager.java
@@ -1,7 +1,0 @@
-package edu.hm2.task3.managers;
-
-import edu.hm2.task3.connections.Connection;
-
-public interface ConnectionManager {
-    Connection getConnection();
-}
