+++ src/main/java/edu/hw2/task1/Constant.java
@@ -1,4 +1,4 @@
-package edu.hm2.task1;
+package edu.hw2.task1;
 
 public record Constant(double value) implements Expr {
