+++ src/main/java/edu/hw2/task1/Expr.java
@@ -1,4 +1,4 @@
-package edu.hm2.task1;
+package edu.hw2.task1;
 
 public sealed interface Expr permits Constant, Negate, Exponent, Addition, Multiplication {
     double evaluate();
