+++ src/main/java/edu/hw2/task2/Square.java
@@ -1,4 +1,4 @@
-package edu.hm2.task2;
+package edu.hw2.task2;
 
 public final class Square extends Rectangle {
