+++ src/main/java/edu/hw2/task3/connections/Connection.java
@@ -1,4 +1,4 @@
-package edu.hm2.task3.connections;
+package edu.hw2.task3.connections;
 
 public interface Connection extends AutoCloseable {
     void execute(String command);
