+++ src/main/java/edu/hw2/task3/connections/StableConnection.java
@@ -1,4 +1,4 @@
-package edu.hm2.task3.connections;
+package edu.hw2.task3.connections;
 
 public class StableConnection implements Connection {
     @Override
