+++ src/main/java/edu/hw2/task3/executors/PopularCommandExecutor.java
@@ -1,7 +1,7 @@
-package edu.hm2.task3.executors;
+package edu.hw2.task3.executors;
 
-import edu.hm2.task3.connections.Connection;
-import edu.hm2.task3.managers.ConnectionManager;
+import edu.hw2.task3.connections.Connection;
+import edu.hw2.task3.managers.ConnectionManager;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
