+++ src/main/java/edu/hw2/task3/managers/ConnectionManager.java
@@ -1,0 +1,7 @@
+package edu.hw2.task3.managers;
+
+import edu.hw2.task3.connections.Connection;
+
+public interface ConnectionManager {
+    Connection getConnection();
+}
