+++ src/main/java/edu/hw2/task3/managers/FaultyConnectionManager.java
@@ -1,7 +1,7 @@
-package edu.hm2.task3.managers;
+package edu.hw2.task3.managers;
 
-import edu.hm2.task3.connections.Connection;
-import edu.hm2.task3.connections.FaultyConnection;
+import edu.hw2.task3.connections.Connection;
+import edu.hw2.task3.connections.FaultyConnection;
 
 public class FaultyConnectionManager implements ConnectionManager {
     @Override
