+++ src/main/java/edu/hw2/task4/CallingInfo.java
@@ -1,4 +1,4 @@
-package edu.hm2.task4;
+package edu.hw2.task4;
 
 public record CallingInfo(String className, String methodName) {
 }
