+++ src/test/java/edu/hw2/Task2Test.java
@@ -1,7 +1,7 @@
 package edu.hw2;
 
-import edu.hm2.task2.Rectangle;
-import edu.hm2.task2.Square;
+import edu.hw2.task2.Rectangle;
+import edu.hw2.task2.Square;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
