+++ src/test/java/edu/hw2/Task4Test.java
@@ -1,9 +1,9 @@
 package edu.hw2;
 
-import edu.hm2.task4.CallingInfo;
+import edu.hw2.task4.CallingInfo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import static edu.hm2.task4.Info.callingInfo;
+import static edu.hw2.task4.Info.callingInfo;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class Task4Test {
